requirements.txt
websockets
httpx
pytest
fastapi
uvicorn
sqlalchemy
pydantic[dotenv]
pyzmq
paho-mqtt